Clean up file organization a bit, separate folder groups a little more cleanly.
